Move everything to screenshot package
