+++ src/main/java/de/retest/web/RecheckSeleniumAdapter.java
@@ -1,6 +1,6 @@
 package de.retest.web;
 
-import static de.retest.web.ScreenshotProviders.shoot;
+import static de.retest.web.screenshot.ScreenshotProviders.shoot;
 
 import java.awt.image.BufferedImage;
 import java.io.IOException;
@@ -26,6 +26,7 @@ import de.retest.recheck.ui.descriptors.RootElement;
 import de.retest.recheck.ui.descriptors.idproviders.RetestIdProvider;
 import de.retest.web.mapping.PathsToWebDataMapping;
 import de.retest.web.screenshot.ScreenshotProvider;
+import de.retest.web.screenshot.ScreenshotProviders;
 import de.retest.web.selenium.UnbreakableDriver;
 import de.retest.web.util.SeleniumWrapperUtil;
 import de.retest.web.util.SeleniumWrapperUtil.WrapperOf;
